+++ adapters.h
@@ -3,9 +3,11 @@
 
 #include "amdgpuadapterhandle.h"
 #include "pciaccess.h"
+#include "structs.h"
 
 class Adapters
 {
+
 private:
 
 public:
@@ -24,39 +26,6 @@ public:
 
     static void PrintAdaptersInfo(AMDGPUAdapterHandle& handle, const std::vector<int>& choosenAdapters, bool useChoosen);
 
-};
-
-struct AdapterIterator
-{
-    const std::vector<int>& adapters;
-    bool allAdapters;
-    int allAdaptersNum;
-    int position;
-
-    AdapterIterator(const std::vector<int>& _adapters, bool _allAdapters, int _allAdaptersNum) :
-        adapters(_adapters), allAdapters(_allAdapters), allAdaptersNum(_allAdaptersNum), position(0)
-    {
-
-    }
-
-    AdapterIterator& operator++()
-    {
-        position++;
-        return *this;
-    }
-
-    operator bool() const
-    {
-        return ( !allAdapters && position < int(adapters.size())) || (allAdapters && position < allAdaptersNum );
-    }
-    bool operator!() const
-    {
-        return ! ( ( !allAdapters && position < int(adapters.size()) ) || (allAdapters && position < allAdaptersNum) );
-    }
-    int operator*() const
-    {
-        return allAdapters ? position : adapters[position];
-    }
 };
 
 #endif /* ADAPTERS_H */
