+++ catalystcrimsonovcparameters.h
@@ -1,0 +1,15 @@
+#ifndef CATALYSTCRIMSONOVCPARAMETERS_H
+#define CATALYSTCRIMSONOVCPARAMETERS_H
+
+class CatalystCrimsonOvcParameters
+{
+
+private:
+
+public:
+
+  static void Set(ADLMainControl& MainControl, const std::vector<int>& ActiveAdapters, const std::vector<OVCParameter>& OvcParams);
+
+};
+
+#endif /* CATALYSTCRIMSONOVCPARAMETERS_H */
