+++ catalystcrimsonprocessing.h
@@ -1,7 +1,8 @@
 #ifndef CATALYSTCRIMSONPARAMETERS_H
 #define CATALYSTCRIMSONPARAMETERS_H
 
-#include "parameters.h"
+#include "catalystcrimsonovcparameters.h"
+#include "structs.h"
 
 class CatalystCrimsonProcessing
 {
@@ -11,7 +12,7 @@ private:
     void checkAdapterList(bool useAdaptersList, std::vector<int> chosenAdapters, std::vector<int> activeAdapters);
 
 public:
-    
+
     void Process(ATIADLHandle Handle_, bool UseAdaptersList, std::vector<int> ChosenAdapters, std::vector<OVCParameter> OvcParameters,
                  bool ChooseAllAdapters, bool PrintVerbose);
